refactor: move UI types to proper namespaces
- Move EditorOverlay, SignaturePanel, ActionMenu from Hex1b.Documents to Hex1b.Widgets
- Move IGutterProvider, DecorationGutterProvider, FoldingGutterProvider, LineNumberGutterProvider from Hex1b to Hex1b.Documents
- Remove redundant self-referencing using statements
- Update all affected consumers with correct using directives
