+++ src/Hex1b/Documents/FoldingGutterProvider.cs
@@ -1,8 +1,7 @@
-using Hex1b.Documents;
 using Hex1b.Layout;
 using Hex1b.Theming;
 
-namespace Hex1b;
+namespace Hex1b.Documents;
 
 /// <summary>
 /// Gutter provider that renders folding indicators (▼/▶) for collapsible regions.
