+++ src/Hex1b/Documents/LineNumberGutterProvider.cs
@@ -1,8 +1,7 @@
-using Hex1b.Documents;
 using Hex1b.Layout;
 using Hex1b.Theming;
 
-namespace Hex1b;
+namespace Hex1b.Documents;
 
 /// <summary>
 /// Built-in gutter provider that renders 1-based line numbers.
