+++ src/Hex1b/LanguageServer/CompletionController.cs
@@ -1,6 +1,7 @@
 using Hex1b.Documents;
 using Hex1b.LanguageServer.Protocol;
 using Hex1b.Theming;
+using Hex1b.Widgets;
 
 namespace Hex1b.LanguageServer;
