+++ src/Hex1b/LanguageServer/LanguageServerDecorationProvider.cs
@@ -2,6 +2,7 @@ using System.Text.Json;
 using Hex1b.Documents;
 using Hex1b.LanguageServer.Protocol;
 using Hex1b.Theming;
+using Hex1b.Widgets;
 
 namespace Hex1b.LanguageServer;
