+++ tests/Hex1b.Tests/OverlayEnhancementTests.cs
@@ -1,5 +1,6 @@
 using Hex1b.Documents;
 using Hex1b.Theming;
+using Hex1b.Widgets;
 
 namespace Hex1b.Tests;
